5398: patron page and remove status.scss
