+++ src/stories/Blocks/status-loans/status-loans.scss
@@ -1,10 +1,15 @@
 .dpl-status-loans {
-  &__column {
-    display: flex;
-    flex-direction: column;
+  display: flex;
+  flex-direction: column;
 
-    @include breakpoint-s {
-      flex-direction: row;
-    }
+  @include breakpoint-s {
+    flex-direction: row;
+    width: 50%;
   }
+}
+
+.dpl-status-loans__container {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
 }
