+++ src/stories/Blocks/status-loans/statusLoans.tsx
@@ -1,8 +1,16 @@
-import { Status, StatusProps } from "../../Library/status/status";
+import {
+  ProgressBar,
+  ProgressBarProps,
+} from "../../Library/progress-bar/progressBar";
 
 export type LinkProps = {
   link: string;
   text: string;
+};
+
+export type StatusProps = {
+  statusBars: ProgressBarProps[];
+  title: string;
 };
 
 export type StatusLoansProps = {
@@ -15,8 +23,9 @@ export type StatusLoansProps = {
 export const StatusLoans = (props: StatusLoansProps) => {
   const { statusBarsData, title: statusBarsTitle, link, bread } = props;
   const { link: url, text: linkText } = link;
+
   return (
-    <div className="dpl-status-loans">
+    <>
       <h1 className="text-header-h1 m-8">{statusBarsTitle}</h1>
       <div className="m-8 text-body-small-regular">
         {`${bread} `}
@@ -24,11 +33,23 @@ export const StatusLoans = (props: StatusLoansProps) => {
           {linkText}
         </a>
       </div>
-      <div className="dpl-status-loans__column">
+      <div className="dpl-status-loans">
         {statusBarsData.map(({ statusBars, title }) => (
-          <Status title={title} statusBars={statusBars} />
+          <div className="dpl-status-loans__container m-8">
+            <h2 className="text-header-h2">{title}</h2>
+            {statusBars.map(
+              ({ title: statusBarTitle, amount, fullAmount, outOf }) => (
+                <ProgressBar
+                  title={statusBarTitle}
+                  fullAmount={fullAmount}
+                  outOf={outOf}
+                  amount={amount}
+                />
+              )
+            )}
+          </div>
         ))}
       </div>
-    </div>
+    </>
   );
 };
