+++ src/stories/Library/Forms/checkbox/checkbox.scss
@@ -11,7 +11,7 @@
 .checkbox__label {
   user-select: none;
   cursor: pointer;
-  padding: 6px 8px;
+  padding: 6px 8px 6px 0;
   overflow: hidden;
   transition: all 0.3s ease;
   display: flex;
