+++ src/stories/Library/Modals/modal-pause/modal-pause.scss
@@ -40,7 +40,7 @@ $modal-pause-container-width: 550px;
     border: 1px solid $c-global-tertiary-1;
     box-sizing: border-box;
     background-color: $c-global-primary;
-    min-width: 230px;
+    min-width: 100%;
     height: 50px;
     padding: 0 10px;
     color: $c-text-secondary-gray;
@@ -78,9 +78,9 @@ $modal-pause-container-width: 550px;
   }
 
   .datepicker-toggle {
+    width: 100%;
     display: inline-block;
     position: relative;
-    width: 18px;
     height: 19px;
   }
 }
