+++ src/stories/Library/dropdown/dropdown.scss
@@ -8,6 +8,14 @@
   }
 }
 
+.dropdown__half-on-desktop {
+  width: 100%;
+
+  @include breakpoint-m {
+    width: 50%;
+  }
+}
+
 .dropdown__select {
   appearance: none;
   -webkit-appearance: none;
