+++ src/stories/Library/patron-info/patron-info.scss
@@ -1,4 +1,4 @@
-.dpl-user-info {
+.dpl-patron-info {
   @extend %color-secondary-gray;
   @extend %bg-global-secondary;
   @extend %pb-32;
