+++ src/stories/Library/patron-page/patron-page.scss
@@ -1,0 +1,8 @@
+.dpl-patron-page {
+  background: $c-global-primary;
+  padding: $s-md;
+
+  @include breakpoint-m {
+    padding: $s-6xl;
+  }
+}
