+++ src/stories/Library/pincode-container/pincode-container.scss
@@ -1,0 +1,16 @@
+.dpl-pincode-container {
+  display: flex;
+  flex-wrap: wrap;
+
+  @include breakpoint-m {
+    flex-wrap: nowrap;
+
+    .dpl-input:first-of-type {
+      margin-right: 8px;
+    }
+
+    .dpl-input:last-of-type {
+      margin-left: 8px;
+    }
+  }
+}
