--- src/stories/Library/status/status.scss
@@ -1,6 +1,0 @@
-.dpl-status {
-  display: flex;
-  flex-direction: column;
-  margin: 8px;
-  flex: 1;
-}
